Fix duplicated route params when calling a handler more than once.
The function produced by COMPILE-REQUIREMENTS for NINGLE-ROUTE was
closing over a single parameter list shared by all invocations of the
route function. The inner functions were pushing requirement
parameters on this list, which left it with parameters from the
previous run on subsequent runs.

Resetting PARAMETERS to NIL before running the inner match functions
would be sufficient, but instead this replaces the somewhat-tricky
lambda-generating code with a more straightforward loop. Match
function lookup is still memoized as before, but instead of iterating
over a list of closures, the generated function just iterates over the
requirement parameters and match functions directly. The only values
closed over are the requirements args and the list of match functions;
hopefully this is un-tricky enough to avoid any future confusion.

There is also a (very) minor performance enhancement to be had here by
accumulating the match parameters once and reversing the list, instead
of producing several intermediate copies of the list with APPEND as in
the previous code.